Move MixinLeavesBlock to more accurate package
